Do not send anything to docker daemon
The docker CLI send the context to the docker daemon in order to build
an image. The Dockerfile is taken from there as well as any file
specified by the ADD and COPY commands.

But this is done in the daemon, so the full context must be sent before
which can degrade the performance of the building.

The .dockerignore file can exclude which files really belong to the
context before sending it.

For now, the file will exclude everything even the Dockerfile (which it
is sent anyway).