Memory and Swap limit
Configured in the compose file the maximum memory that the container can
use. Disabled the swap.
Created a test to verify this. A simple program that will consume an
arbitrary amount of memory: if the resource limits of docker works as
expected, the program that would try to consume more than this limit
should be killed automatically (actually, I think that there are not any
warranty that this and only this program will be killed, the kernel may
kill other process in the container).